refactored buildlog type
